refactor(components): introduce DateFilter interface and update related components
- Added DateFilter interface to encapsulate date filter display entries and status labels.
- Updated ArticlePageHeader, CollectionCard, and related components to utilize the new DateFilter structure for improved type safety and maintainability.
- Removed deprecated imports and streamlined date filter handling across components.
