+++ packages/components/src/interfaces/internal/CollectionCard.ts
@@ -1,39 +1,14 @@
 import type { ImageProps } from "~/interfaces"
 import type { FormattedDate, IsomerSiteProps, TagGroup } from "~/types"
-import type { DateFilterStatusId } from "~/types/constants"
+
+import type { DateFilterDisplayEntry } from "./DateFilter"
 
 // NOTE: one entry per date-type filter the item has a raw value for — used
 // for filter matching (see getFilteredItems' range-overlap check).
-export interface DateFilterValue {
+interface DateFilterValue {
   id: string
   date: string
   endDate?: string
-}
-
-export interface DateFilterStatusLabel {
-  id: DateFilterStatusId
-  label: string
-}
-
-// NOTE: server-precomputed display fields for a date filter (label + formatted
-// date text + admin status labels). Live status is derived on the client.
-export interface DateFilterDisplayEntry {
-  id: string
-  label: string
-  dateText: string
-  date: string
-  endDate?: string
-  statusLabels: DateFilterStatusLabel[]
-}
-
-// NOTE: fully resolved card display entry, including live status — only
-// produced on the client (see DateFilterStatusClient).
-export interface DateFilterCard {
-  id: string
-  label: string
-  status: DateFilterStatusId
-  statusLabel: string
-  dateText: string
 }
 
 interface FileDetails {
