+++ packages/components/src/templates/next/components/internal/ArticlePageHeader/ArticlePageHeader.stories.tsx
@@ -1,6 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react-vite"
 import type { ArticlePageHeaderProps } from "~/interfaces"
-import type { DateFilterStatusLabel } from "~/interfaces/internal/CollectionCard"
 import { DATE_FILTER_STATUS_ID } from "~/types/constants"
 
 import { ArticlePageHeader } from "./ArticlePageHeader"
@@ -14,13 +13,13 @@ const daysFromNow = (days: number) => {
   return toDateString(date)
 }
 
-const EVENT_DATE_STATUS_LABELS: DateFilterStatusLabel[] = [
+const EVENT_DATE_STATUS_LABELS = [
   { id: DATE_FILTER_STATUS_ID.Ended, label: "Event ended" },
   { id: DATE_FILTER_STATUS_ID.Ongoing, label: "Ongoing" },
   { id: DATE_FILTER_STATUS_ID.Upcoming, label: "Upcoming" },
 ]
 
-const REGISTRATION_DEADLINE_STATUS_LABELS: DateFilterStatusLabel[] = [
+const REGISTRATION_DEADLINE_STATUS_LABELS = [
   { id: DATE_FILTER_STATUS_ID.Ended, label: "Registration closed" },
   { id: DATE_FILTER_STATUS_ID.Ongoing, label: "Registration open" },
   { id: DATE_FILTER_STATUS_ID.Upcoming, label: "Registration upcoming" },
